refactor: move stat tests into separate file
